requirements.txt
flask==2.0.1
line-bot-sdk==1.16.0
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
gunicorn==20.1.0 